Add safer interface for A* path with callback
Instead of forcing the user to deal with the unsafe code directly by having to
implement the callback passed to `TCOD_path_new_using_function` (which involves
having to cast from `*void` to whatever type you're using for your user_data and
also making sure the the path object doesn't outlive user_data) we let the users
pass in a closure and handle all the nitty-gritty ourselves.

A similar treatments awaits the Map-based version (the path object should own
the Map else we're looking at another memory safety issues when the Map goes
away before the path does) and the Dijkstra variants.

We need tests for this.