Run tests with Travis
Because `cargo test` builds the examples as well, Rust needs to be able
to link against libtcod.so.

Since it's not packaged in Ubuntu (which is what our Travis tests run
on), we download the release binaries from the official website and copy
them to where all the other system libraries are.