Use unboxed closures for path finding
This needs a bit of a hackery to cast the closure trait object to C *void
user_data and back, but it's worth it.

We want to use closures that capture by value because that lets us create the
path finding objects in a function and just solves a bunch of lifetime issues
the previous implementation had.