Update path finding examples to unboxed closures
The switch to unboxed closures itself is tiny (just opt into the feature and add
`&mut:` in front of the closure), but to demonstrate the benefit, we're now
creating the closure & the path finding object in a separate function.

Rust's by-ref closures are not allowed to escape the stack frame in which they
were defined but the by-val ones can. Which is exactly what we want in this
situation.