Don't use transmute_copy
I was using it to avoid passing ownership of the transmutees, but it's more
dangerous because it's not doing any size checks. This has already bitten me.

So this switches to met::transmute. That means the code is more verbose because
we need to explicitly take back the ownership of anything we allocate on the
heap, but we get the safety of size checking.