Add terminal.png
This allows running examples using `cargo run --example example_name` with a
recent Cargo.

The non-console examples worked before as well, but whenever you call
`Console::init_root` it either expects a font to be set or looks for
`terminal.png` in the current directory.

We're relying on the fallback to keep the examples more focused.