Bring back the `Color` type to `tcod-rs`
`tcod-sys` still needs its own `TCOD_color_t` for its FFI signatures, but we
want to define some methods on `tcod::Color`. And we can't do that if it's not
defined in the same crate as the methods.

Same if we wanted to implement the operator traits (+, *, etc.).

We could shove all that implementation to `tcod-sys` but that's not how they
should be separated.

So this creates a new struct with the same internal representation as
`TCOD_color_t` which we then transmute back and forth when calling out to FFI.

The one downside of this is that we cannot return the static colour definitions
directly anymore, so this just sets the right values to the colour names.

But the good news is that most of libtcod's Color stuff is now exported.