Added unsafe to Console::con
It returns a mutable pointer without requiring &mut self. For the implementation's sake this is the ideal behaviour, but this still means that calling it is an unsafe operation. Unfortunately there's no way to remove this from the public interface without crippling console::blit, as this requires the underlying native representation.