Removed some exposed native types
Since this is a safe wrapper it is better not to provide methods that could potentially cause unsafe behavior, especially when traits can solve this problem wonderfully: the unsafe methods were removed from Color's API. They were moved into the implementations of the AsNative and FromNative traits, both of which the user will explicitly have to include if they want to use the native types. As an additional safety net, there's also an unsafe qualifier on the traits' respective methods.