Use `AsRef<Path>` for fns that take a fs path
This let's us pass a str and Path to `RootInitializer::font` for example.

We get back the ergonomics of accepting `&str`, but if a user has a Path, they
can still just pass it in.

The situation here is complicated by the fact that we have to store the value in
the `RootInitializer` struct, which can be done either by taking a Reference to
`AsPath` or by saving it in a Box.

I've opted for the latter since the struct is short-lived and it will only be
done once. Otherwise, we'd have to put the str and Path behind an reference:

Root::initializer().font("terminal.png").init();

vs.

Root::initializer().font(&"terminal.png").init();