Implement rustc_serialize for Color
This is an optional feature (because it depends on an external crate people may
possibly not want to include). We have to do the impls here in tcod because our
users can't reasonably do it themself: both the Encodable/Decadable traits and
the Color struct come from third-party crates and Rust prevents implementing a
trait for a type when you don't control either.

We can do the same treatment for other serialisation solutions in the future. I
looked at Serde but the version on Crates.io is different from the current HEAD,
so let's wait until it gets a bit more stable.