Simplify the unsafe callback code in path finding
We don't have to cast the callback closure to a trait object and
therefore store its fat pointer and pass that as libtcod's user data. By
making `c_path_callback` generic over the closure, we can pass the
pointer to the closure as user data directly.

This saves us an extra indirection, makes the code simpler and gets rid
of `mem::transmute`.

I've tested this on the examples and ran them through valgrind. Astar is
clean as before, Dijkstra is showing memory issues, but again, that's
same as before.