Replace external `Time` crate with `std::time`
`std::time::Duration` is stable as of Rust 1.3. Using it simplifies our
dependencies and should make interoperability with other code easier.

There is a couple of downsides:

First, the API of the Duration in std is smaller than the one in the
time crate so we need to e.g. do our own conversion to total seconds and
milliseconds elapsed. I'm hoping this will get better in the future or
will be something the time crate will adopt while using
std::time::Duration for the underlying data.

Second, this is a breaking change because it no longer works with the
older (<=1.2) compilers.