Fix MSVC targets
I'm not 100% sure what the underlying issue was, but there have been
warnings from the Visual Studio compiler and linker about our generated
`*.o` and `*.a` files and how we were including them and I think they
may have ended up being ignored.

So I've changed the MSVC branch to pass the C files to the final
compiler directly and just build everything in one command. This, in
addition to some compile switches to make it work seems to do the trick.