Update Serde and deprecate rustc-serialize
rustc-serialize is going to be deprecated soon:

https://internals.rust-lang.org/t/rust-release-milestone-predictions/4591

In the meantime, Serde is now perfectly usable on stable.

So this deprecates rustc-serialize and will be removed from the a future
release. It also brings our Serde bindings up-to-date and replace the
manual implementation with a conditional custom derive.

[breaking-change]

We've replaced the "serde" feature with "serialization". This is because
the feature names and optional crates can't conflict and we need both
`serde` and `serde_derive`.

So if you want to use tcod with the `Serialize`/`Deserialize` traits,
put this in your `Cargo.toml`:

    tcod = { version = "*", features = ["serialization"] }

Closes #235